feat:75[動画視聴画面] タグの削除APIのアクセス制限

+++ server/services/bookmark/destroy.ts
@@ -1,7 +1,8 @@
-import type { FastifySchema } from "fastify";
+import type { FastifyRequest, FastifySchema } from "fastify";
 import authUser from "$server/auth/authUser";
 import deleteBookmark from "$server/utils/bookmark/deleteBookmark";
 import { BookmarkParams } from "$server/validators/bookmarkParams";
+import findBookmark from "$server/utils/bookmark/findBookmark";
 
 export const destroySchema: FastifySchema = {
   summary: "ブックマークの削除",
@@ -10,6 +11,7 @@ export const destroySchema: FastifySchema = {
   response: {
     204: { type: "null", description: "成功" },
     400: {},
+    403: {},
   },
 };
 
@@ -17,8 +19,25 @@ export const destroyHooks = {
   auth: [authUser],
 };
 
-export async function destroy({ params }: { params: BookmarkParams }) {
+export async function destroy({
+  session,
+  params,
+}: FastifyRequest<{ Params: BookmarkParams }>) {
   const { id } = params;
+  const bookmark = await findBookmark(id);
+
+  if (!bookmark) {
+    return {
+      status: 400,
+    };
+  }
+
+  if (bookmark.userId !== session.user.id) {
+    return {
+      status: 403,
+    };
+  }
+
   const deleted = await deleteBookmark({
     id,
   });
